+++ spec/gruf/error_spec.rb
@@ -62,7 +62,8 @@ describe Gruf::Error do
       end
 
       context 'with a very large error message' do
-        let(:message) { SecureRandom.hex(10000) }
+        let(:message) { SecureRandom.hex(described_class::MAX_METADATA_SIZE + 100) }
+
         it 'should log the original error and replace the outgoing error with a new one' do
           # expect(Gruf.logger).to receive(:warn)
           expect(subject).to be_a(described_class)
@@ -152,7 +153,7 @@ describe Gruf::Error do
           fe = resp.field_errors.first
           expect(fe).to_not be_a(Gruf::FieldError)
           expect(fe).to_not be_a(Gruf::FieldError)
-        end.to raise_error(Gruf::Client::Error)
+        end.to raise_error(Gruf::Client::Errors::InvalidArgument)
       end
     end
 
@@ -161,7 +162,7 @@ describe Gruf::Error do
         client = build_client
         expect do
           client.call(:GetContextualFieldErrorFail, id: 1)
-        end.to raise_error(Gruf::Client::Error)
+        end.to raise_error(Gruf::Client::Errors::InvalidArgument)
 
         expect do
           resp = client.call(:GetContextualFieldErrorFail, id: 2)
@@ -172,7 +173,6 @@ describe Gruf::Error do
 
     context 'with multiple calls, with multiple having field errors' do
       it 'should not aggregate errors across calls', run_thing_server: true do
-
         ts = []
         10.times do
           ts << Thread.new do
@@ -182,7 +182,7 @@ describe Gruf::Error do
               resp = client.call(:GetContextualFieldErrorFail, id: 1)
               expect(resp.error).to be_a(Gruf::ErrorHeader)
               expect(resp.field_errors.count).to eq(1)
-            end.to raise_error(Gruf::Client::Error)
+            end.to raise_error(Gruf::Client::Errors::InvalidArgument)
           end
         end
         ts.map(&:join)
@@ -213,7 +213,7 @@ describe Gruf::Error do
         resp = client.call(:GetException, id: 1)
         expect(resp.error).to be_a(Gruf::ErrorHeader)
         expect(resp.error_code).to eq :internal
-      end.to raise_error(Gruf::Client::Error)
+      end.to raise_error(Gruf::Client::Errors::Internal)
     end
   end
 end